requirements/local.txt
asgiref==3.8.1
Django==4.1
django-model-utils==5.0.0
face-recognition-models==0.3.0
mysql==0.0.3
mysqlclient==2.2.5
pillow==11.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2